+++ jkube-kit/enricher/api/src/test/java/org/eclipse/jkube/kit/enricher/api/BaseEnricherTest.java
@@ -36,10 +36,9 @@ import java.util.Properties;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.eclipse.jkube.kit.enricher.api.BaseEnricher.getNamespace;
-import static org.mockito.Answers.RETURNS_DEEP_STUBS;
-import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.spy;
 import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
 
 class BaseEnricherTest {
   private BaseEnricher baseEnricher;
@@ -435,5 +434,37 @@ class BaseEnricherTest {
 
     // Then
     assertThat(config).isNotNull();
+  }
+
+  @Test
+  void shouldUseLegacyJKubePrefix_whenPropertyProvided_thenReturnTrue() {
+    // Given
+    when(context.getProperty("jkube.useLegacyJKubePrefix")).thenReturn("true");
+    // When
+    boolean value = baseEnricher.shouldUseLegacyJKubePrefix();
+    // Then
+    assertThat(value).isTrue();
+  }
+
+  @Test
+  void shouldUseLegacyJKubePrefix_whenResourceConfigProvided_thenReturnTrue() {
+    // Given
+    baseEnricher = new BaseEnricher(context.toBuilder()
+        .resources(ResourceConfig.builder()
+            .useLegacyJKubePrefix(true)
+            .build())
+        .build(), "test-enricher");
+    // When
+    boolean value = baseEnricher.shouldUseLegacyJKubePrefix();
+    // Then
+    assertThat(value).isTrue();
+  }
+
+  @Test
+  void shouldUseLegacyJKubePrefix_whenNothingProvided_thenReturnFalse() {
+    // Given + When
+    boolean value = baseEnricher.shouldUseLegacyJKubePrefix();
+    // Then
+    assertThat(value).isFalse();
   }
 }